[Refactor] Usage of the new Configuration way
