+++ lib/map_layers/map_leyers.dart
@@ -57,7 +57,7 @@ class MapLayer extends MapTileProvider {
   final MapLayerIds mapLayerId;
   final String mapKey;
 
-  MapLayer(this.mapLayerId, this.mapKey) : super();
+  MapLayer(this.mapLayerId, {this.mapKey}) : super();
 
   @override
   List<LayerOptions> buildTileLayerOptions() {
